Encapsulate behaviour in classes
Current solution made was not thread safe and it was not possible to send
notifications using two different configurations (well it was possble, but not
reliable). Current solution for sure works for APNS, GCM was not tested yet. This
also prooves, that test suite for poshmeup should be extended ASAP.